perf(listener): move cache.ts outof listener module
